requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
pip==20.0.2
pkg-resources==0.0.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
setuptools==44.0.0
sqlparse==0.4.3
